requirements.txt
Flask
Flask-SQLAlchemy
flask_cors
flask_restful
python-dotenv
pytest
psycopg2-binary
Flask-Migrate
flask_jwt_extended
bcrypt
PyJWT 
requests
mercadopago
qrcode
Pillow
openpyxl
brcode
pixqrcode
loguru

